array, array_buffer etc. now use usize

+++ core/engine/src/builtins/array/array_iterator.rs
@@ -30,7 +30,7 @@ use boa_gc::{Finalize, Trace};
 #[derive(Debug, Clone, Finalize, Trace, JsData)]
 pub(crate) struct ArrayIterator {
     array: JsObject,
-    next_index: u64,
+    next_index: usize,
     #[unsafe_ignore_trace]
     kind: PropertyNameKind,
     done: bool,
